requirements-ocr.txt
Pillow>=10.0.0
PyYAML>=6.0
pytesseract>=0.3.10
